+++ clang-tools-extra/test/clang-tidy/checkers/performance/bitwise-operation-on-bool.cpp
@@ -1,0 +1,202 @@
+// RUN: %check_clang_tidy %s performance-bitwise-operation-on-bool %t
+
+bool& normal() {
+    int a = 100, b = 200;
+
+    a | b;
+    a & b;
+    a |= b;
+    a &= b;
+
+    a bitor b;
+    a bitand b;
+    a or_eq b;
+    a and_eq b;
+
+    static bool st = false;
+    return st;
+}
+
+void bad() {
+    bool a = true, b = false;
+    a | b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a || b;
+    a & b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a && b;
+    a |= b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a || b;
+    a &= b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a && b;
+}
+
+void bad_volatile_bool() {
+    bool a = true;
+    volatile bool b = false;
+    a | b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES-NOT: :[[@LINE-2]]:{{.*}}: note: FIX-IT applied suggested code changes
+    a & b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES-NOT: :[[@LINE-2]]:{{.*}}: note: FIX-IT applied suggested code changes
+
+    a |= b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES-NOT: :[[@LINE-2]]:{{.*}}: note: FIX-IT applied suggested code changes
+    a &= b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES-NOT: :[[@LINE-2]]:{{.*}}: note: FIX-IT applied suggested code changes
+}
+
+void bad_nontraditional() {
+    bool a = true, b = false;
+    a bitor b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a or b;
+    a bitand b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a and b;
+    a or_eq b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a or b;
+    a and_eq b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a and b;
+}
+
+void bad_with_priors() {
+    bool a = false, b = true, c = true;
+    a && b | c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:10: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a && (b || c);
+    a && b & c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:12: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a && b && c;
+    a || b & c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:12: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a || b && c;
+    a || b | c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:12: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a || b || c;
+    b | c && a;
+    // CHECK-MESSAGES: :[[@LINE-1]]:5: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: (b || c) && a;
+}
+
+void bad_with_priors2() {
+    bool a = false, b = true, c = true;
+    a ^ b & c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:9: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a ^ (b && c);
+    a | b & c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES: :[[@LINE-2]]:11: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a || b && c;
+    b & c ^ a;
+    // CHECK-MESSAGES: :[[@LINE-1]]:5: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: (b && c) ^ a;
+}
+
+void bad_with_priors_compound() {
+    bool a = false, b = true, c = true;
+    a &= b || c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a && (b || c);
+    a |= b || c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a || b || c;
+    a &= b && c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a && b && c;
+    a |= b && c;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a || b && c;
+}
+
+void bad_with_priors_compound2() {
+    // TODO: ^ and |
+}
+
+void bad_no_fixit() {
+    bool b = false;
+    normal() |= b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:14: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES-NOT: :[[@LINE-2]]:{{.*}}: note: FIX-IT applied suggested code changes
+    normal() &= b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:14: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-MESSAGES-NOT: :[[@LINE-2]]:{{.*}}: note: FIX-IT applied suggested code changes
+}
+
+#define MY_OR |
+#define MY_AND &
+#define MY_OR_ASSIGN |=
+#define MY_AND_ASSIGN &=
+
+#define MY_OR_FUNC(a, b) ((a) | (b))
+#define MY_AND_FUNC(a, b) ((a) & (b))
+#define MY_OR_ASSIGN_FUNC(a, b) ((a) |= (b))
+#define MY_AND_ASSIGN_FUNC(a, b) ((a) &= (b))
+
+void bad_in_macro() {
+    bool a = true, b = false;
+
+    // TODO: same for braces
+
+    a MY_OR b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a || b;
+    a MY_AND b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a && b;
+    a MY_OR_ASSIGN b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a || b;
+    a MY_AND_ASSIGN b;
+    // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use logical operator instead of bitwise one for bool [performance-bitwise-operation-on-bool]
+    // CHECK-FIXES: a = a && b;
+
+    // TODO: MY_OR_FUNC(a, b);
+    // TODO: MY_AND_FUNC(a, b);
+    // TODO: MY_OR_ASSIGN_FUNC(a, b);
+    // TODO: MY_AND_ASSIGN_FUNC(a, b);
+}
+
+// TODO: all the same tests as in math-parentheses check
+void bad_in_macro___() {
+#if 0
+    const std::string Input = R"cc(
+#define M(a,b) (true & a) * (true | b)
+    int f() { return M(false, false); }
+    )cc";
+#endif
+// TODO: implement this(make sure no fix were provided)
+#if 0
+    const std::string Input = R"cc(
+#define M(a,b) (true a) * (true b)
+    int f() { return M(& false, | false); }
+    )cc";
+    const std::string Expected = R"cc(
+#define M(a,b) (true a) * (true b)
+    int f() { return M(&& false, || false); }
+    )cc";
+#endif
+// TODO: implement this
+}
+
+template<typename T>
+void good_in_unreachable_template(T a, T b) {
+    a | b;
+    a & b;
+    a |= b;
+    a &= b;
+}
+
+// TODO: test for in template
+// TODO: test for type in typedef
+// TODO: test for expressions in parentheses
+
+
+